+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # CHANGELOG
 
+## 1.73.6 (April 7, 2025)
+
+ENHANCEMENTS:
+
+* make v5 assume role duration configurable [GH-6616]
+
+## 1.73.5 (April 3, 2025)
+
+* **New Data Source:**
+  + `huaweicloud_networking_secgroup_tags` [GH-6611]
+
+BUG FIXES:
+
+* resource/huaweicloud_cfw_black_white_list: fix address issue [GH-6608]
+
 ## 1.73.4 (April 2, 2025)
 
 * **New Resource:**